requirements.txt
absl-py==2.2.2
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cvzone==1.6.1
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.5.1
gast==0.6.0
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.32.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==2.15.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyMatting==1.1.13
pyparsing==3.2.3
pyreadline3==3.5.4
pyspellchecker==0.8.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
tifffile==2025.3.30
timm==1.0.15
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.14.1
